+++ subscript.js
@@ -6,7 +6,7 @@ import './feature/string.js'
 import './feature/access.js'
 import './feature/group.js'
 import './feature/prefix.js'
-import './feature/postfix.js'
+import './feature/inc.js'
 import './feature/mult.js'
 import './feature/add.js'
 import './feature/bitwise.js'
