+++ test/test.js
@@ -1,4 +1,4 @@
-import './tree.js'
+import './parse.js'
 import './subscript.js'
 import './jsep.js'
 // import './perf.js'
